Autogenerate lzfse-sys bindings (and bump major verions)
This makes the bindings match the types of the header (which is nice,
since the header uses u8s rather than c chars). It also gets the
lzfse_encode_scratch_size function which was previously missed.

Also, add functions to use a scratch space rather than always passing a
null scratch pointer.

This changes the types of lzfse-sys, so this is a breaking change for
that crate, incremented to v2.0. This only adds new apis to `lzfse`, so
only bumped the minor version.